Move dictionary and file utils over to dict-common
