+++ link-grammar/Makefile.am
@@ -83,6 +83,7 @@ liblink_grammar_la_SOURCES =        \
 	count.c                          \
 	dict-common/build-disjuncts.c    \
 	dict-common/dict-common.c        \
+	dict-common/dict-utils.c         \
 	dict-common/idiom.c              \
 	dict-common/print-dict.c         \
 	dict-common/regex-morph.c        \
@@ -129,6 +130,7 @@ liblink_grammar_la_SOURCES =        \
 	dict-common/dict-api.h           \
 	dict-common/dict-common.h        \
 	dict-common/dict-structures.h    \
+	dict-common/dict-utils.h         \
 	dict-common/idiom.h              \
 	dict-common/regex-morph.h        \
 	dict-file/read-dict.h            \
